+++ libraries/exoplayer_dash/src/main/java/androidx/media3/exoplayer/dash/DefaultDashChunkSource.java
@@ -20,11 +20,13 @@ import static java.lang.Math.min;
 
 import android.net.Uri;
 import android.os.SystemClock;
+import android.util.Pair;
 import androidx.annotation.CheckResult;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
 import androidx.media3.common.Format;
 import androidx.media3.common.util.UnstableApi;
+import androidx.media3.common.util.UriUtil;
 import androidx.media3.common.util.Util;
 import androidx.media3.datasource.DataSource;
 import androidx.media3.datasource.DataSpec;
@@ -729,12 +731,17 @@ public class DefaultDashChunkSource implements DashChunkSource {
               ? 0
               : DataSpec.FLAG_MIGHT_NOT_USE_FULL_NETWORK_SPEED;
       ImmutableMap<@CmcdConfiguration.HeaderKey String, String> httpRequestHeaders =
-          cmcdHeadersFactory == null
-              ? ImmutableMap.of()
-              : cmcdHeadersFactory
-                  .setChunkDurationUs(endTimeUs - startTimeUs)
-                  .setObjectType(CmcdHeadersFactory.getObjectType(trackSelection))
-                  .createHttpRequestHeaders();
+          ImmutableMap.of();
+      if (cmcdHeadersFactory != null) {
+        Pair<String, String> nextObjectAndRangeRequest =
+            getNextObjectAndRangeRequest(firstSegmentNum, segmentUri, representationHolder);
+        cmcdHeadersFactory
+            .setChunkDurationUs(endTimeUs - startTimeUs)
+            .setObjectType(CmcdHeadersFactory.getObjectType(trackSelection))
+            .setNextObjectRequest(nextObjectAndRangeRequest.first)
+            .setNextRangeRequest(nextObjectAndRangeRequest.second);
+        httpRequestHeaders = cmcdHeadersFactory.createHttpRequestHeaders();
+      }
       DataSpec dataSpec =
           DashUtil.buildDataSpec(
               representation,
@@ -779,12 +786,17 @@ public class DefaultDashChunkSource implements DashChunkSource {
               ? 0
               : DataSpec.FLAG_MIGHT_NOT_USE_FULL_NETWORK_SPEED;
       ImmutableMap<@CmcdConfiguration.HeaderKey String, String> httpRequestHeaders =
-          cmcdHeadersFactory == null
-              ? ImmutableMap.of()
-              : cmcdHeadersFactory
-                  .setChunkDurationUs(endTimeUs - startTimeUs)
-                  .setObjectType(CmcdHeadersFactory.getObjectType(trackSelection))
-                  .createHttpRequestHeaders();
+          ImmutableMap.of();
+      if (cmcdHeadersFactory != null) {
+        Pair<String, String> nextObjectAndRangeRequest =
+            getNextObjectAndRangeRequest(segmentNum, segmentUri, representationHolder);
+        cmcdHeadersFactory
+            .setChunkDurationUs(endTimeUs - startTimeUs)
+            .setObjectType(CmcdHeadersFactory.getObjectType(trackSelection))
+            .setNextObjectRequest(nextObjectAndRangeRequest.first)
+            .setNextRangeRequest(nextObjectAndRangeRequest.second);
+        httpRequestHeaders = cmcdHeadersFactory.createHttpRequestHeaders();
+      }
       DataSpec dataSpec =
           DashUtil.buildDataSpec(
               representation,
@@ -808,6 +820,23 @@ public class DefaultDashChunkSource implements DashChunkSource {
           sampleOffsetUs,
           representationHolder.chunkExtractor);
     }
+  }
+
+  private Pair<String, String> getNextObjectAndRangeRequest(
+      long segmentNum, RangedUri segmentUri, RepresentationHolder representationHolder) {
+    if (segmentNum + 1 < representationHolder.getSegmentCount()) {
+      RangedUri nextSegmentUri = representationHolder.getSegmentUrl(segmentNum + 1);
+      Uri uri = segmentUri.resolveUri(representationHolder.selectedBaseUrl.url);
+      Uri nextUri = nextSegmentUri.resolveUri(representationHolder.selectedBaseUrl.url);
+      String nextObjectRequest = UriUtil.getRelativePath(uri, nextUri);
+
+      String nextRangeRequest = nextSegmentUri.start + "-";
+      if (nextSegmentUri.length != C.LENGTH_UNSET) {
+        nextRangeRequest += (nextSegmentUri.start + nextSegmentUri.length);
+      }
+      return new Pair<>(nextObjectRequest, nextRangeRequest);
+    }
+    return new Pair<>(null, null);
   }
 
   private RepresentationHolder updateSelectedBaseUrl(int trackIndex) {